Fixed shadow acne, and normal maps for rough and image textures.

+++ Utils/vec2.h
@@ -7,43 +7,43 @@
 
 class vec2 {
 public:
-    float e[2];
+    double e[2];
 
 public:
     // Constructors
     vec2();
-    vec2(float x, float y);
+    vec2(double x, double y);
 
     // Getters
-    inline float getX() const { return e[0]; }
-    inline float getY() const { return e[1]; }
-    void setX(float val);
-    void setY(float val);
+    inline double getX() const { return e[0]; }
+    inline double getY() const { return e[1]; }
+    void setX(double val);
+    void setY(double val);
 
     // Def vector operators
     inline const vec2& operator+() const { return *this; }
     inline vec2 operator-() const { return vec2(-e[0], -e[1]); }
-    inline float operator[](int i) const { return e[i]; }
-    inline float& operator[](int i) { return e[i]; }
+    inline double operator[](int i) const { return e[i]; }
+    inline double& operator[](int i) { return e[i]; }
 
     // Scalar operations
     inline vec2& operator+=(const vec2& v2);
     inline vec2& operator-=(const vec2& v2);
     inline vec2& operator*=(const vec2& v2);
     inline vec2& operator/=(const vec2& v2);
-    inline vec2& operator*=(const float t);
-    inline vec2& operator/=(const float t);
+    inline vec2& operator*=(const double t);
+    inline vec2& operator/=(const double t);
 
     // Basic math operations
-    inline float length() const { return sqrt(e[0] * e[0] + e[1] * e[1]); }
-    inline float lengthSquared() const { return e[0] * e[0] + e[1] * e[1]; }
+    inline double length() const { return sqrt(e[0] * e[0] + e[1] * e[1]); }
+    inline double lengthSquared() const { return e[0] * e[0] + e[1] * e[1]; }
     void makeUnitVec();
 };
 
 // Constructor
 vec2::vec2() {}
 
-vec2::vec2(float x, float y) {
+vec2::vec2(double x, double y) {
     e[0] = x;
     e[1] = y;
 }
@@ -60,7 +60,7 @@ inline std::ostream& operator<<(std::ostream& os, vec2& t) {
 }
 
 inline void vec2::makeUnitVec() {
-    float k = 1.0 / sqrt(e[0] * e[0] + e[1] * e[1]);
+    double k = 1.0 / sqrt(e[0] * e[0] + e[1] * e[1]);
     e[0] *= k;
     e[1] *= k;
 }
@@ -78,16 +78,16 @@ inline vec2 operator*(const vec2& v1, const vec2& v2) {
 inline vec2 operator/(const vec2& v1, const vec2& v2) {
     return vec2(v1.e[0] / v2.e[0], v1.e[1] / v2.e[1]);
 }
-inline vec2 operator*(float t, const vec2& v) {
+inline vec2 operator*(double t, const vec2& v) {
     return vec2(t * v.e[0], t * v.e[1]);
 }
-inline vec2 operator*(const vec2& v, float t) {
+inline vec2 operator*(const vec2& v, double t) {
     return vec2(t * v.e[0], t * v.e[1]);
 }
-inline vec2 operator/(float t, const vec2& v) {
+inline vec2 operator/(double t, const vec2& v) {
     return vec2(t / v.e[0], t / v.e[1]);
 }
-inline vec2 operator/(const vec2& v, float t) {
+inline vec2 operator/(const vec2& v, double t) {
     return vec2(v.e[0] / t, v.e[1] / t);
 }
 
@@ -112,21 +112,21 @@ inline vec2& vec2::operator/=(const vec2& v) {
     e[1] /= v.e[1];
     return *this;
 }
-inline vec2& vec2::operator*=(const float t) {
+inline vec2& vec2::operator*=(const double t) {
     e[0] *= t;
     e[1] *= t;
     return *this;
 }
 
-inline vec2& vec2::operator/=(const float t) {
-    float k = 1.0 / t;
+inline vec2& vec2::operator/=(const double t) {
+    double k = 1.0 / t;
     e[0] /= k;
     e[1] /= k;
     return *this;
 }
 
 // Dot products
-inline float dot(const vec2& v1, const vec2& v2) {
+inline double dot(const vec2& v1, const vec2& v2) {
     return v1[0] * v2[0] + v1[1] * v2[1];
 }
