+++ Utils/vec3.h
@@ -7,39 +7,39 @@
 
 class vec3 {
 public:
-	float e[3];
+	double e[3];
 
 public:
 	// Constructors
 	vec3();
-	vec3(float x, float y, float z);
+	vec3(double x, double y, double z);
 
 	// Getters
-	inline float getX() const { return e[0]; }
-	inline float getY() const { return e[1]; }
-	inline float getZ() const { return e[2]; }
-	inline void setX(float val) { e[0] = val; }
-	inline void setY(float val) { e[1] = val; }
-	inline void setZ(float val) { e[2] = val; }
+	inline double getX() const { return e[0]; }
+	inline double getY() const { return e[1]; }
+	inline double getZ() const { return e[2]; }
+	inline void setX(double val) { e[0] = val; }
+	inline void setY(double val) { e[1] = val; }
+	inline void setZ(double val) { e[2] = val; }
 
 	// Def vector operators
 	inline const vec3& operator+() const { return *this; }
 	inline vec3 operator-() const { return vec3(-e[0], -e[1], -e[2]); }
-	inline float operator[](int i) const { return e[i]; }
-	inline float& operator[](int i) { return e[i]; }
+	inline double operator[](int i) const { return e[i]; }
+	inline double& operator[](int i) { return e[i]; }
 
 	// Scalar operations
 	inline vec3& operator+=(const vec3& v2);
 	inline vec3& operator-=(const vec3& v2);
 	inline vec3& operator*=(const vec3& v2);
 	inline vec3& operator/=(const vec3& v2);
-	inline vec3& operator*=(const float t);
-	//inline vec3& operator/=(const float t);
+	inline vec3& operator*=(const double t);
+	//inline vec3& operator/=(const double t);
 	inline vec3& operator/=(const double t) { return *this *= 1 / t; }
 
 	// Basic math operations
-	inline float length() const { return sqrt(e[0] * e[0] + e[1] * e[1] + e[2] * e[2]); }
-	inline float lengthSquared() const { return e[0] * e[0] + e[1] * e[1] + e[2] * e[2]; }
+	inline double length() const { return sqrt(e[0] * e[0] + e[1] * e[1] + e[2] * e[2]); }
+	inline double lengthSquared() const { return e[0] * e[0] + e[1] * e[1] + e[2] * e[2]; }
 	void makeUnitVec();
 	bool nearZero() const;
 	vec3 fract() const;
@@ -50,7 +50,7 @@ public:
 // Constructor
 vec3::vec3() {}
 
-vec3::vec3(float x, float y, float z) {
+vec3::vec3(double x, double y, double z) {
 	e[0] = x;
 	e[1] = y;
 	e[2] = z;
@@ -69,7 +69,7 @@ inline std::ostream& operator<<(std::ostream& os, vec3& t) {
 
 inline void vec3::makeUnitVec() {
 	// Get 1 / len and multiply all the components by the new magnitude
-	float k = 1.0 / sqrt(e[0] * e[0] + e[1] * e[1] + e[2] * e[2]);
+	double k = 1.0 / sqrt(e[0] * e[0] + e[1] * e[1] + e[2] * e[2]);
 	e[0] *= k; 
 	e[1] *= k; 
 	e[2] *= k;
@@ -88,16 +88,16 @@ inline vec3 operator*(const vec3& v1, const vec3& v2) {
 inline vec3 operator/(const vec3& v1, const vec3& v2) {
 	return vec3(v1.e[0] / v2.e[0], v1.e[1] / v2.e[1], v1.e[2] / v2.e[2]);
 }
-inline vec3 operator*(float t, const vec3& v) {
+inline vec3 operator*(double t, const vec3& v) {
 	return vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
-inline vec3 operator*(const vec3& v, float t) {
+inline vec3 operator*(const vec3& v, double t) {
 	return vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
-inline vec3 operator/(float t, const vec3& v) {
+inline vec3 operator/(double t, const vec3& v) {
 	return vec3(t / v.e[0], t / v.e[1], t / v.e[2]);
 }
-inline vec3 operator/(const vec3& v, float t) {
+inline vec3 operator/(const vec3& v, double t) {
 	return vec3(v.e[0] / t, v.e[1] / t, v.e[2] / t);
 }
 
@@ -126,7 +126,7 @@ inline vec3& vec3::operator/=(const vec3& v) {
 	e[2] /= v.e[2];
 	return *this;
 }
-inline vec3& vec3::operator*=(const float t) {
+inline vec3& vec3::operator*=(const double t) {
 	e[0] *= t;
 	e[1] *= t;
 	e[2] *= t;
@@ -134,8 +134,8 @@ inline vec3& vec3::operator*=(const float t) {
 }
 
 /*
-inline vec3& vec3::operator/=(const float t) {
-	float k = 1.0 / t;
+inline vec3& vec3::operator/=(const double t) {
+	double k = 1.0 / t;
 	e[0] /= k;
 	e[1] /= k;
 	e[2] /= k;
@@ -143,7 +143,7 @@ inline vec3& vec3::operator/=(const float t) {
 }*/
 
 // Dot and Cross products
-inline float dot(const vec3 & v1, const vec3 & v2) {
+inline double dot(const vec3 & v1, const vec3 & v2) {
 	return v1[0] * v2[0] + v1[1] * v2[1] + v1[2] * v2[2];
 }
 
@@ -162,7 +162,7 @@ inline vec3 unitVector(vec3 v) {
 vec3 randInUnitSphere() {
 	vec3 p;
 	do {
-		p = 2.9 * vec3((float)rand() / RAND_MAX, (float)rand() / RAND_MAX, (float)rand() / RAND_MAX) - vec3(1, 1, 1);
+		p = 2.9 * vec3((double)rand() / RAND_MAX, (double)rand() / RAND_MAX, (double)rand() / RAND_MAX) - vec3(1, 1, 1);
 	} while (p.lengthSquared() >= 1.0);
 	return p;
 }
@@ -172,10 +172,10 @@ vec3 reflect(const vec3& vector, const vec3& normal) {
 	return vector - 2 * dot(vector, normal) * normal;
 }
 
-bool refract(const vec3& vector, const vec3& normal, float ior, vec3& refractedVector) {
+bool refract(const vec3& vector, const vec3& normal, double ior, vec3& refractedVector) {
 	vec3 uv = unitVector(vector);
-	float dt = dot(uv, normal);
-	float discriminant = 1.0 - ior * ior * (1 - dt * dt);
+	double dt = dot(uv, normal);
+	double discriminant = 1.0 - ior * ior * (1 - dt * dt);
 	if (discriminant > 0) {
 		refractedVector = ior * (uv - normal * dt) - normal * sqrt(discriminant);
 		return true;
@@ -185,8 +185,8 @@ bool refract(const vec3& vector, const vec3& normal, float ior, vec3& refractedV
 }
 
 // Shlick approx
-float shlick(float cosine, float ior) {
-	float r0 = (1 - ior) / (1 + ior);
+double shlick(double cosine, double ior) {
+	double r0 = (1 - ior) / (1 + ior);
 	r0 = r0 * r0;
 	return r0 + (1 - r0) * (pow((1 - cosine), 5));
 }
